refactor(property): move buildFormData logic to a separate utility function
